Use types with explicit sizes
Focusing on the diskimg library this time, which deals with a lot of
filesystem structures that have specific widths.

This is still a bit lax in places, e.g. using "long" for lengths.
Should either specify a bit width or use di_off_t.

Also, added "override" keyword where appropriate.

Also, bumped library version to 5.0.0.

+++ diskimg/GenericFD.h
@@ -133,7 +133,7 @@ public:
      * be seeked to their initial positions.  "length" bytes will be copied.
      */
     static DIError CopyFile(GenericFD* pDst, GenericFD* pSrc, di_off_t length,
-            unsigned long* pCRC = NULL);
+            uint32_t* pCRC = NULL);
 
 protected:
     GenericFD& operator=(const GenericFD&);
